fix review-auto : fallback "jamais d'ecran vide" + portabilite + tests
CONTEXTE : suite a /code-review (xhigh, 9 angles) sur la PR #1. Tri des
findings (plusieurs refutes par gardes existantes) ; fixes pertinents
appliques a la racine, pas de rustine.

FICHIERS CLES :
  - src/core/director.{hpp,cpp} : nouveau resolvePlayable() — si la cible (un
    locuteur) n'a aucune scene jouable (pool vide / speaker introuvable),
    fallback sur le plus fort puis sur le plan large ; si rien n'est jouable,
    on PURGE la memoisation pour re-tenter au tick suivant (plus de blocage sur
    une cible morte, plus jamais d'ecran vide quand quelqu'un parle).
  - src/core/weighted_pick.hpp : cumul des poids en long long (long = 32 bits
    sur Windows / 64 sur Linux -> comportement identique multi-OS, pas de
    depassement, pas de warning CI macOS).
  - CMakeLists.txt : find_package(Qt6 REQUIRED ...) (coherent avec libobs ;
    echec clair si Qt absent au lieu d'une erreur de link obscure).
  - scripts/install-local.bat : refuse de s'executer si OBS est ouvert (evite
    une install corrompue : data effacee + DLL verrouillee).
  - .gitignore : ignore CMakeUserPresets.json, compile_commands.json, *.user
    (chemins absolus machine-specifiques).
  - tests/test_core.cpp : 2 tests trompeurs corriges (contexte B 'wide' et
    contexte C 'last vs wide' reellement exerces) + couverture ajoutee
    (rafraichissement temps-max via RNG controlable, anti-scintillement Single,
    fallback pool vide, sans-scene-ni-large, forceScene->reprise auto,
    round-trip JSON complet tous champs, fromJson leve sur JSON invalide).

VERIF (fichiers reels) : coeur 25 cas / 131 assertions / SUCCESS, 0 warning ;
plugin streamdirector.dll recompile 0 erreur / 0 warning.

ZONES A RISQUE : aucune nouvelle. Notes Run 2 : edge 'current = plan large
force non-standard' (modele owner/wide a enrichir), dette setConfig (reset
detecteurs), raffinements plugin-main (thread guard, i18n titre, message UI).

COMPAT : none (projet solo, PR pas encore mergee).

REF : /code-review xhigh PR #1 ; changelog 2026-05-30 (PR #1 + code-review).

+++ src/core/weighted_pick.hpp
@@ -12,7 +12,10 @@ namespace sd::core {
 // Les poids <= 0 sont ignores. Renvoie nullptr si aucune option eligible.
 template <typename T>
 const T* weightedPick(const std::vector<std::pair<T, int>>& options, double r) {
-    long total = 0;
+    // long long (et pas long) : long fait 32 bits sur Windows (LLP64) et 64 sur
+    // Linux/macOS (LP64) — on fixe 64 bits partout pour un comportement identique
+    // multi-OS et eviter tout depassement sur de gros cumuls de poids.
+    long long total = 0;
     for (const auto& opt : options) {
         if (opt.second > 0) {
             total += opt.second;
